Split form content into an iframe
Temporary solution to bootstrap breaking forms downstream from the menu.
